adjusted some stuff

+++ LIBFT/ft_strlcpy.c
@@ -1,8 +1,10 @@
 
-int ft_strlcpy(char *dst, const char *src, int size)
+#include <stddef.h>
+
+size_t ft_strlcpy(char *dst, const char *src, size_t size)
 {
-    int i;
-    int src_len;
+    size_t i;
+    size_t src_len;
 
     i = 0;
     src_len = 0;
